+++ typescript/dbos-nextjs-starter/.dbosignore
@@ -14,8 +14,8 @@
 /coverage
 
 # next.js
-**/.next/**
-/out/
+.next/
+out/
 
 # production
 /build
