+++ typescript/dbos-nextjs-starter/README.md
@@ -1,88 +1,18 @@
-# DBOS Background Job
+# Welcome to DBOS!
 
-This is a [DBOS app](https://docs.dbos.dev/) bootstrapped with `npx @dbos-inc/create`, using [Express.js](https://expressjs.com/) and [Knex](https://docs.dbos.dev/typescript/tutorials/orms/using-knex) to interact with postgres.
+This is a template app built with DBOS and Express.
 
-## Getting Started
+### Getting Started
 
-Before you can launch your app, you need a database.
-DBOS works with any Postgres database, but to make things easier, we've provided a script that starts a Docker Postgres container and creates a database.
-Run:
-
-```bash
-node start_postgres_docker.js
-```
-
-If successful, the script should print `Database started successfully!`.
-
-```
-
-## Production build
-
-In production, instead of using `nodemon`, the following separate steps should be used to build, run database setup, and start the app.
-
-```bash
-npm run build
-```
-
-Then, run a schema migration to create some tables:
-
-```bash
-npx dbos migrate
-```
-
-If successful, the migration should print `Migration successful!`.
-
-Finally, run the app:
-
-```bash
-npm run start
-```
-
-To see that it's working, visit this URL in your browser: [`http://localhost:3000/`](http://localhost:3000/).
-
-Click on the "Start Background Job" button.
-
-You should this message: `Your background task has completed step 0 of 9.`
-As the background job completes steps, the step counter should go up by one!
-
-Click on "Crash the application" button.
-
-The step counter stops increasing.
-On the command line, you will see that the application has stopped.
-Start the application again.
-```bash
-npm run start
-```
-In the browser, you will see that the execution of steps resumes where it left off and eventually completes.
-
-Congratulations! You just run a DBOS application.
-
-## The application
-
-- In `src/action/dbosWorkflow.tsx`, the DBOS workflow is created. When called, it increments a counter in the database and returns a greeting.
-
-- The code below in the same file, launches the dbos runtime. This is required for DBOS runtime to start. Do not remove the code
-```
-if (process.env.NEXT_PHASE !== "phase-production-build") {
-    await DBOS.launch();
-}
-```
-- The workflow for the background job is called by the called the GET method in app/tasks/route.ts.
-
-- The GET is called by the component in src/components/BackGroundTask.tsx. It calls the route /tasks.
-
-- The component is called from the main UI page.tsx.
-
-- This is how the DBOS workflow is wired to the NextJs application.
+To get started building, edit `src/main.ts`.
+Then, commit your changes and visit the [cloud console](https://console.dbos.dev/applications) to redeploy it from GitHub!
 
 
- DBOS will wrap all routes with an [OpenTelemetry](https://opentelemetry.io/) tracing middleware and tie HTTP traces to DBOS workflow traces.
+<details>
+<summary><strong>Deploying via the DBOS Cloud CLI</strong></summary>
 
-To add more functionality to this application, modify `src/operations.ts`. If you used `npm run dev`, it will automatically rebuild and restart.
-
-## Running in DBOS Cloud
-
-To deploy this app to DBOS Cloud, first install the DBOS Cloud CLI (example with [npm](https://www.npmjs.com/)):
+You can also deploy this app via the DBOS Cloud CLI.
+Install it globally with this command:
 
 ```shell
 npm i -g @dbos-inc/dbos-cloud
@@ -93,8 +23,63 @@ Then, run this command to deploy your app:
 ```shell
 dbos-cloud app deploy
 ```
+</details>
 
-## Next Steps
+### Developing Locally
 
-- For a detailed tutorial, check out our [programming quickstart](https://docs.dbos.dev/typescript/programming-guide).
-- To learn more about DBOS, take a look at [our documentation](https://docs.dbos.dev/) or our [source code](https://github.com/dbos-inc).
+To run this app locally, you need to connect it to a Postgres database.
+You can use a DBOS Cloud database, a Docker container, or a local Postgres installation.
+
+<details>
+<summary><strong>Connecting to a DBOS Cloud Postgres database</strong></summary>
+
+> You must have already deployed an app to DBOS Cloud before using the DBOS Cloud Postgres database.
+
+First install the DBOS Cloud CLI globally:
+
+```shell
+npm i -g @dbos-inc/dbos-cloud
+```
+
+Then set a password for your DBOS Cloud database:
+
+```shell
+dbos-cloud db reset-password
+```
+
+Then connect your local app to your cloud database. When prompted, enter the password you just set.
+
+```shell
+dbos-cloud db local
+```
+</details>
+
+<details>
+<summary><strong>Starting Postgres with Docker</strong></summary>
+
+If you have Docker, you can start a Postgres database locally with:
+
+```shell
+export PGPASSWORD=dbos
+node start_postgres_docker.js
+```
+</details>
+
+
+After connecting to Postgres, install dependencies and start your app:
+
+```shell
+npm install
+npm run build
+npx dbos migrate
+npm run start
+```
+
+or simply run in the dev mode:
+
+```shell
+npm install
+npm run dev
+```
+
+Visit [`http://localhost:3000`](http://localhost:3000) to see your app!
