+++ typescript/dbos-nextjs-starter/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "hello-nextjs",
+  "name": "dbos-nextjs-starter",
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev --turbopack",
+    "dev": "next dev",
     "build": "next build",
     "start": "next start",
     "lint": "next lint"
