--- typescript/dbos-nextjs-starter/src/actions/crash.ts
@@ -1,5 +1,0 @@
-"use server";
-
-export async function crash() {
-    process.exit(1);
-}
