+++ typescript/dbos-nextjs-starter/src/app/page.tsx
@@ -2,9 +2,7 @@
 
 import Head from 'next/head';
 import { useState, useEffect, useRef } from 'react';
-import { lastStep } from '@/actions/last_step';
-import { startBackgroundTask } from '@/actions/background';
-import { crash } from '@/actions/crash';
+import { crash, lastStep, startBackgroundTask } from '@/actions/background';
 
 export default function Page() {
   const [currentId, setCurrentId] = useState<string | null>(null);
@@ -12,6 +10,7 @@ export default function Page() {
   const [reconnecting, setReconnecting] = useState(false);
   const currentIdRef = useRef<string | null>(currentId);
 
+  // Periodically check the progress of the background task
   useEffect(() => {
     currentIdRef.current = currentId; // Keep the ref updated with the current value
   }, [currentId]);
@@ -31,7 +30,6 @@ export default function Page() {
         setStatus(`Your background task has completed <b>${step} of 10</b> steps`);
         if (step === '10') {
           setCurrentId(null);
-          clearInterval(interval); // Stop the interval
         }
         setReconnecting(false);
       } catch (error) {
@@ -71,11 +69,6 @@ export default function Page() {
     <>
       <Head>
         <title>Welcome to DBOS!</title>
-        <link
-          rel="icon"
-          href="https://dbos-blog-posts.s3.us-west-1.amazonaws.com/live-demo/favicon.ico"
-          type="image/x-icon"
-        />
       </Head>
       <main className="font-sans text-gray-800 p-6 max-w-2xl mx-auto">
         <h1 className="text-xl font-semibold mb-4">Welcome to DBOS!</h1>
@@ -84,25 +77,9 @@ export default function Page() {
             id="reconnecting"
             className="mb-4 bg-yellow-100 p-2 rounded-md text-center flex items-center justify-center gap-2"
           >
-            <svg
-              className="animate-spin h-5 w-5 text-yellow-700"
-              xmlns="http://www.w3.org/2000/svg"
-              fill="none"
-              viewBox="0 0 24 24"
-            >
-              <circle
-                className="opacity-25"
-                cx="12"
-                cy="12"
-                r="10"
-                stroke="currentColor"
-                strokeWidth="4"
-              ></circle>
-              <path
-                className="opacity-75"
-                fill="currentColor"
-                d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
-              ></path>
+              <svg className="animate-spin h-5 w-5 text-yellow-700" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
+                <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" stroke-width="4"></circle>
+                <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
             </svg>
             <span className="text-yellow-700">Reconnecting...</span>
           </div>
@@ -143,14 +120,14 @@ export default function Page() {
         <p className="mb-4">
           To start building your own crashproof application, access your source code from the{' '}
           <a
-            href="https://console.dbos.dev/applications/dbos-node-starter"
+            href="https://console.dbos.dev/applications/dbos-nextjs-starter"
             target="_blank"
             rel="noopener noreferrer"
             className="text-blue-600 hover:underline"
           >
             cloud console
           </a>
-          , edit <code className="bg-gray-100 px-1 rounded">src/main.ts</code>, then redeploy your app.
+          , edit <code className="bg-gray-100 px-1 rounded">src/dbos/operations.ts</code>, then redeploy your app.
         </p>
         <p className="mb-4">
           To learn how to build crashproof apps with DBOS, check out the{' '}
